requirements.txt
flask
Flask-CORS
opencv-contrib-python-headless
python-dotenv
ultralytics
requests
waitress
gunicorn==20.0.4
numpy
pillow
